Bug fix for account editor on Android
Emulator starts offline (navigator.onLine is false). And trying to go online was not working.
EditAccountPage was throwing an error when one tried to go online.
EditAccountPage is always watching the app.offlineTracker and before the fix tried to access its model when a toggle takes place.
And it was failing if the model was undefined.

Now it won't fail if it doesn't have a model.

NB:It would be better for it to start/stop watching the app.offlineTracker as it comes into view/leaves view but that change is more involved.

Also sync screen won't be shown if there is nothing to sync.