force.js copied from https://github.com/ccoenraets/forcejs/blob/a9348b90a046e12aac9b56864359c24e5b246abe/force.js
This is the unmodified force.js
In an upcoming PR, I'll make changes to it (to work with the network plugin) and will change sample / template / test apps to use it instead of forcetk.
I want the upcoming PR diff to show the changes I made to force.js (instead of just showing it as all new).